moved function to forteachersmodule
