+++ app.py
@@ -1412,32 +1412,6 @@ def hour_of_code(level, program_id=None):
         ))
 
 
-def migrate_quizzes_parsons_tabs(customizations, parsons_hidden, quizzes_hidden):
-    """If the puzzles/quizzes were not migrated yet which is possible if the teacher didn't tweak
-        the class customizations, if this is the case, we need to add them if possible."""
-    migrated = customizations.get("quiz_parsons_tabs_migrated")
-    if not migrated and customizations.get("sorted_adventures"):
-        for level, sorted_adventures in customizations["sorted_adventures"].items():
-            last_two_adv_names = [adv["name"] for adv in sorted_adventures[-2:]]
-            parson_in_level = "parsons" in last_two_adv_names
-            quiz_in_level = "quiz" in last_two_adv_names
-            # In some levels, we don't need quiz/parsons
-            level_accepts_parsons = "parsons" in hedy_content.ADVENTURE_ORDER_PER_LEVEL[int(level)]
-            level_accepts_quiz = "quiz" in hedy_content.ADVENTURE_ORDER_PER_LEVEL[int(level)]
-            if not parson_in_level and not parsons_hidden and level_accepts_parsons:
-                sorted_adventures.append(
-                    {"name": "parsons", "from_teacher": False})
-
-            if not quiz_in_level and not quizzes_hidden and level_accepts_quiz:
-                sorted_adventures.append(
-                    {"name": "quiz", "from_teacher": False})
-            # Need to reorder, for instance, in case parsons was hidden and the other was not.
-            for_teachers.ForTeachersModule.reorder_adventures(sorted_adventures)
-
-        # Mark current customization as being migrated so that we don't do this step next time.
-        customizations["quiz_parsons_tabs_migrated"] = 1
-        DATABASE.update_class_customizations(customizations)
-
 # routing to index.html
 
 
@@ -1611,7 +1585,7 @@ def index(level, program_id):
     quizzes_hidden = 'other_settings' in customizations and 'hide_quiz' in customizations['other_settings']
 
     if customizations:
-        migrate_quizzes_parsons_tabs(customizations, parsons_hidden, quizzes_hidden)
+        for_teachers.ForTeachersModule.migrate_quizzes_parsons_tabs(customizations, parsons_hidden, quizzes_hidden)
 
     parsons_in_level = True
     quiz_in_level = True
