+++ website/for_teachers.py
@@ -28,7 +28,6 @@ from website.auth import (
 from .achievements import Achievements
 from .database import Database
 from .website_module import WebsiteModule, route
-from app import migrate_quizzes_parsons_tabs
 
 SLIDES = collections.defaultdict(hedy_content.NoSuchSlides)
 for lang in hedy_content.ALL_LANGUAGES.keys():
@@ -543,6 +542,33 @@ class ForTeachersModule(WebsiteModule):
                               available_adventures=available_adventures,
                               class_id=session['class_id'])
 
+    @staticmethod
+    def migrate_quizzes_parsons_tabs(customizations, parsons_hidden, quizzes_hidden):
+        """If the puzzles/quizzes were not migrated yet which is possible if the teacher didn't tweak
+            the class customizations, if this is the case, we need to add them if possible."""
+        migrated = customizations.get("quiz_parsons_tabs_migrated")
+        if not migrated and customizations.get("sorted_adventures"):
+            for level, sorted_adventures in customizations["sorted_adventures"].items():
+                last_two_adv_names = [adv["name"] for adv in sorted_adventures[-2:]]
+                parson_in_level = "parsons" in last_two_adv_names
+                quiz_in_level = "quiz" in last_two_adv_names
+                # In some levels, we don't need quiz/parsons
+                level_accepts_parsons = "parsons" in hedy_content.ADVENTURE_ORDER_PER_LEVEL[int(level)]
+                level_accepts_quiz = "quiz" in hedy_content.ADVENTURE_ORDER_PER_LEVEL[int(level)]
+                if not parson_in_level and not parsons_hidden and level_accepts_parsons:
+                    sorted_adventures.append(
+                        {"name": "parsons", "from_teacher": False})
+
+                if not quiz_in_level and not quizzes_hidden and level_accepts_quiz:
+                    sorted_adventures.append(
+                        {"name": "quiz", "from_teacher": False})
+                # Need to reorder, for instance, in case parsons was hidden and the other was not.
+                ForTeachersModule.reorder_adventures(sorted_adventures)
+
+            # Mark current customization as being migrated so that we don't do this step next time.
+            customizations["quiz_parsons_tabs_migrated"] = 1
+            Database.update_class_customizations(Database, customizations)
+
     def get_class_info(self, user, class_id, get_customizations=False):
         if hedy_content.Adventures(g.lang).has_adventures():
             default_adventures = hedy_content.Adventures(g.lang).get_adventure_keyname_name_levels()
@@ -623,7 +649,7 @@ class ForTeachersModule(WebsiteModule):
             self.db.update_class_customizations(customizations)
 
         if get_customizations:
-            migrate_quizzes_parsons_tabs(customizations, parsons_hidden, quizzes_hidden)
+            self.migrate_quizzes_parsons_tabs(customizations, parsons_hidden, quizzes_hidden)
 
         for level, sorted_adventures in customizations['sorted_adventures'].items():
             for adventure in sorted_adventures:
